Remove eslint check for default props, since we're using TS to do props typing
